Simplify, and drive all the box titles & title-links from the Content API
Wendy says it's ok for the all the box titles to be driven by the
Content-API, which means that the 'News' title becomes 'UK news'. This
is ok (say editorial) because Guardian iPad Edition *is* The UK Paper...
we don't publish editorially different editions of the paper in other
parts of the world.